Long is 32-bit and actually use 32-bit must.
I tried generating libc using posix but ran into issues

+++ src/lib_c/asmjs-unknown-emscripten/c/stddef.cr
@@ -1,3 +1,3 @@
 lib LibC
-  alias SizeT = ULong
+  alias SizeT = UInt
 end
